requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic[email]
python-dotenv